% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gbif_completeness.R
\name{clean_gbif_completeness}
\alias{clean_gbif_completeness}
\title{clean_gbif_completeness}
\usage{
clean_gbif_completeness(gbif_df, iucn_shp, taxa_list, method = "zizka",
  crs = NULL)
}
\arguments{
\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{input}{df or raster}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
