% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{completeness}
\usage{
completeness(gbif_df, raster_iucn, crs = "longlat")
}
\arguments{
\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{input}{df or raster}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
