% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_analysis.R
\name{temporal_analysis}
\alias{temporal_analysis}
\title{completeness}
\usage{
temporal_analysis(input, raster_input)
}
\arguments{
\item{input}{df or raster}

\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
