% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_analysis_to_raster.R
\name{temporal_analysis_to_raster}
\alias{temporal_analysis_to_raster}
\title{temporal_analysis_to_raster}
\usage{
temporal_analysis_to_raster(input, from = NULL, to = NULL,
  ratio_low = NULL, ratio_high = NULL, var_2 = NULL,
  remove_ratio_low = NULL, remove_ratio_high = NULL)
}
\arguments{
\item{input}{df or raster}

\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
