% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_per_cell.R
\name{visits_per_cell}
\alias{visits_per_cell}
\title{visits_per_cell}
\usage{
visits_per_cell(input, min_n_visits = 5, extent_object)
}
\arguments{
\item{input}{df or raster}

\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
